fixed stupid file organization
